Project: take in update logic
