+++ src/main/java/seedu/address/logic/parser/SetStatusCommandParser.java
@@ -14,8 +14,8 @@ import java.util.Optional;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.SetStatusCommand;
-import seedu.address.logic.commands.SetStatusCommand.SetStatusDescriptor;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.tag.Project.SetStatusDescriptor;
 
 /**
  * Parses input arguments and creates a new EditCommand object
